rag-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pdfplumber==0.11.4
chromadb==0.5.20
sentence-transformers==3.3.1
httpx==0.27.2
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
prometheus-fastapi-instrumentator==7.0.0
